+++ protoc_plugin/lib/src/client_generator.dart
@@ -10,7 +10,29 @@ class ClientApiGenerator {
   final String className;
   final Set<String> usedMethodNames = {...reservedMemberNames};
 
-  ClientApiGenerator(this.service, Set<String> usedNames)
+  /// Tag of `FileDescriptorProto.service`.
+  static const _fileDescriptorServiceTag = 6;
+
+  /// Tag of `ServiceDescriptorProto.method`.
+  static const _serviceDescriptorMethodTag = 2;
+
+  /// Index of the service in `FileDescriptorProto.service` repeated field.
+  final int _repeatedFieldIndex;
+
+  List<int> get _serviceDescriptorPath => [
+        ...service.fileGen.fieldPath,
+        _fileDescriptorServiceTag,
+        _repeatedFieldIndex
+      ];
+
+  List<int> _methodDescriptorPath(int methodRepeatedFieldIndex) => [
+        ..._serviceDescriptorPath,
+        _serviceDescriptorMethodTag,
+        methodRepeatedFieldIndex
+      ];
+
+  ClientApiGenerator(
+      this.service, Set<String> usedNames, this._repeatedFieldIndex)
       : className = disambiguateName(
             avoidInitialUnderscore(service._descriptor.name),
             usedNames,
@@ -20,6 +42,10 @@ class ClientApiGenerator {
   String get _clientType => '$protobufImportPrefix.RpcClient';
 
   void generate(IndentingWriter out) {
+    final commentBlock = service.fileGen.commentBlock(_serviceDescriptorPath);
+    if (commentBlock != null) {
+      out.println(commentBlock);
+    }
     if (service._descriptor.options.deprecated) {
       out.println(
           '@$coreImportPrefix.Deprecated(\'This service is deprecated\')');
@@ -29,15 +55,16 @@ class ClientApiGenerator {
       out.println('${className}Api(this._client);');
       out.println();
 
-      for (final m in service._descriptor.method) {
-        generateMethod(out, m);
+      for (var i = 0; i < service._descriptor.method.length; i++) {
+        generateMethod(out, service._descriptor.method[i], i);
       }
     });
     out.println();
   }
 
   // Subclasses can override this.
-  void generateMethod(IndentingWriter out, MethodDescriptorProto m) {
+  void generateMethod(IndentingWriter out, MethodDescriptorProto m,
+      int methodRepeatedFieldIndex) {
     final methodName = disambiguateName(
         avoidInitialUnderscore(service._methodName(m.name)),
         usedMethodNames,
@@ -45,6 +72,11 @@ class ClientApiGenerator {
     final inputType = service._getDartClassName(m.inputType, forMainFile: true);
     final outputType =
         service._getDartClassName(m.outputType, forMainFile: true);
+    final commentBlock = service.fileGen
+        .commentBlock(_methodDescriptorPath(methodRepeatedFieldIndex));
+    if (commentBlock != null) {
+      out.println(commentBlock);
+    }
     if (m.options.deprecated) {
       out.println(
           '@$coreImportPrefix.Deprecated(\'This method is deprecated\')');